Refactored bears as well
